pyproject.toml
[tool.poetry]
name = "guarddog"
description = "GuardDog is a CLI tool to Identify malicious PyPI packages"
authors = ["Ellen Wang", "Christophe Tafani-Dereeper"]
license = "Apache-2.0"
readme = "pypi.rst"
repository = "https://github.com/DataDog/guarddog"
version = "0.0.0"

[tool.poetry.scripts]
guarddog = "guarddog.cli:cli"

[tool.poetry.dependencies]
python = ">=3.10,<4"
docker = "==6.0.0b1"
# NOTE: Before https://github.com/returntocorp/semgrep/issues/6631 is addressed, we can't seem to upgrade past 0.112.1
semgrep = "==0.112.1"
requests = "==2.28.1"
tqdm = "==4.64.0"
python-dotenv = "==0.20.0"
python-dateutil = "==2.8.2"
attrs = "==21.4.0"
boltons = "==21.0.0"
bracex = "==2.3.post1"
charset-normalizer = "==2.1.0"
click = "==8.1.3"
click-option-group = "==0.5.3"
colorama = "==0.4.5"
defusedxml = "==0.7.1"
face = "==20.1.1"
glom = "==22.1.0"
idna = "==3.3"
jsonschema = "==4.9.1"
pyparsing = "==3.0.9"
python-lsp-jsonrpc = "==1.0.0"
six = "==1.16.0"
typing-extensions = "==4.3.0"
ujson = "==5.4.0"
urllib3 = "==1.26.11"
wcmatch = "==8.4"
websocket-client = "==1.3.3"
mypy-extensions = "==0.4.3"
pathspec = "==0.9.0"
platformdirs = "==2.5.2"
tomli = "==2.0.1"
pathos = "^0.2.9"
pytest = "^7.1.2"
flake8 = "^5.0.4"
python-whois = "^0.8.0"
termcolor = "^2.1.0"
setuptools = "^65.6.3"
tarsafe = "^0.0.4"
semantic-version = "^2.10.0"
pyyaml = "^6.0"
packaging = "21"
    
[tool.poetry.dev-dependencies]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
testpaths = [
    "tests/analyzer/metadata",
    "tests/core"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

